refactor: extract and organise HomePageTabs folder
