+++ src/components/GenericPagesComponents/AiTab/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./AiTab";
