+++ src/components/GenericPagesComponents/CurriculumTab/CurriculumTab.stories.tsx
