+++ src/components/GenericPagesComponents/CurriculumTab/CurriculumTab.tsx
