+++ src/components/GenericPagesComponents/CurriculumTab/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./CurriculumTab";
