+++ src/components/GenericPagesComponents/PupilTab/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./PupilTab";
